requirements.txt
streamlit==1.25.0
pandas==1.5.3
numpy==1.24.2
matplotlib==3.6.2
networkx==3.4.2
pyvis==0.3.2
graphviz==0.20.1
scipy==1.10.0
requests==2.28.1
plotly==5.13.1
pydot==1.4.2
openai ==1.55.3
streamlit-flow-component ==1.5.0
pydantic
streamlit-option-menu
scikit-learn

